% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModule.R
\name{addDestroyers}
\alias{addDestroyers}
\title{Add shiny.destroy Code to Module}
\usage{
addDestroyers(module)
}
\arguments{
\item{module}{The function call to `moduleServer`}
}
\value{
An updated version of `module`, where the `{shiny.destroy}`
code has been added.
}
\description{
For a given `moduleServer` call, add the code required for `{shiny.destroy}`
to work. This will involve creating the observer list to the user session,
and adds all observers within the list.
}
