% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_module.R
\name{destroyableModuleUI}
\alias{destroyableModuleUI}
\alias{destroyableModuleServer}
\title{Create Dynamic Module}
\usage{
destroyableModuleUI(id, ui, wrapper = div)

destroyableModuleServer(id, module, session = getDefaultReactiveDomain())
}
\description{
Create Dynamic Module
}
