% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroyModule.R
\name{destroyModule}
\alias{destroyModule}
\title{Destroy Shiny Module}
\usage{
destroyModule(id = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{The module namespace ID}

\item{session}{The shiny session, by default it is `shiny::getDefaultReactiveDomain()`}
}
\description{
Given the namespace
}
