% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroyModule.R
\name{destroyModuleServer}
\alias{destroyModuleServer}
\title{Destroy the Server Component of a Shiny Module}
\usage{
destroyModuleServer(id, session = getDefaultReactiveDomain())
}
\description{
Destroy the Server Component of a Shiny Module
}
