% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroyModule.R
\name{destroyModuleUI}
\alias{destroyModuleUI}
\title{Destroy the UI Component of a Shiny Module}
\usage{
destroyModuleUI(id, session = getDefaultReactiveDomain())
}
\description{
Destroy the UI Component of a Shiny Module
}
