% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModule.R
\name{isSpecifiedFunction}
\alias{isSpecifiedFunction}
\title{Check if Function Call is relevant function}
\usage{
isSpecifiedFunction(fn_call, fns)
}
\arguments{
\item{fn_call}{A function call}

\item{fns}{A character vector of functions to compare the function call against}
}
\value{
A logical value stating whether or not the function call is in the collection.
}
\description{
A short description...
}
