% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeInput.R
\name{removeInput}
\alias{removeInput}
\title{Remove Input from Shiny Session}
\usage{
removeInput(id, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Input value name}

\item{session}{The Shiny session to remove the input from}
}
\value{
An invisible `TRUE` value confirming that the input has been removed.
}
\description{
The removal of the named input in a shiny session.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(shiny)
library(shiny.destroy)

ui <- fluidPage(
  numericInput("number", "Select number:", 5, 1, 10),
  p("Selected number:", textOutput("number_out", inline = TRUE)),
  actionButton("delete", "Remove input")
)

server <- function(input, output, session) {
  output$number_out <- renderText(input$number)

  observeEvent(input$delete, removeInput("number"))
}

shinyApp(ui, server)
\dontshow{\}) # examplesIf}
}
