% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{runDestroyExample}
\alias{runDestroyExample}
\title{Run `shiny.destroy` example}
\usage{
runDestroyExample(example = NA, ...)
}
\arguments{
\item{example}{The name of the example to run, or \code{NA} (the default) to
list the available examples.}

\item{...}{Additional parameters sent to `shiny::runExample`}
}
\description{
To see how the `shiny.destroy` works, examples are provided within
the package.
}
\details{
The following examples are available:

\describe{
\item{01_boxes}{A simple application where the "create" button will
load a simple box with a "destroy" button. This highlights the full
removal of the module when the button is pressed.}
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
runDestroyExample("01_boxes")
\dontshow{\}) # examplesIf}
}
