% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.destroy.R
\docType{package}
\name{shiny.destroy-package}
\alias{shiny.destroy}
\alias{shiny.destroy-package}
\title{shiny.destroy: Create Destroyable Modules in 'Shiny'}
\description{
This package enables the complete removal of various 'Shiny' components, such as inputs, outputs and modules. It also aids in the removal of observers that have been created in dynmically created modules.
}
\author{
\strong{Maintainer}: Ashley Baldry \email{arbaldry91@gmail.com}

}
