% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_module.R
\name{destroyableModuleUI}
\alias{destroyableModuleUI}
\alias{makeModuleServerDestroyable}
\title{Create Dynamic Module}
\usage{
destroyableModuleUI(id, ui, wrapper = div)

makeModuleServerDestroyable(module)
}
\description{
Create Dynamic Module
}
