% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModule.R
\name{makeModuleUIDestroyable}
\alias{makeModuleUIDestroyable}
\alias{makeModuleServerDestroyable}
\title{Create Destroyable Module}
\usage{
makeModuleUIDestroyable(module_fn, wrapper = shiny::div)

makeModuleServerDestroyable(module_fn)
}
\arguments{
\item{module_fn}{The server-side part of the module}

\item{wrapper}{If the module is a `shiny.tag.list`, then an HTML tag
will be wrapped by an HTML tag so that a shiny.destroy attribute
can be attached}
}
\description{
A short description...
}
\examples{
library(shiny)

# UI
basicModuleUI <- function(id) {
  ns <- NS(id)
  actionButton("click")
}

makeModuleUIDestroyable(basicModuleUI)

# Server-side
basicMoudleServer <- function(id) {
  moduleServer(id, function(input, output, session) {
    rv <- reactiveVal()
    observeEvent(input$click, rv(input$click))
  })
}

makeModuleServerDestroyable(basicMoudleServer)

# Shiny Application
ui <- fluidPage(
  destroyableModuleUI(id = "test"),
  actionButton("destroy", "Destroy module"),
  textOutput("reactive_value")
)

server <- function(input, output, session) {
  top_rv <- reactiveVal()

  reactive_value <- destroyableModuleServer("test")
  observeEvent(reactive_value(), top_rv(reactive_value()))

  output$reactive_value <- renderText(top_rv())

  observeEvent(input$destroy, destroyModule("test"))
}

}
