% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.destroy.R
\docType{package}
\name{shiny.destroy-package}
\alias{shiny.destroy}
\alias{shiny.destroy-package}
\title{shiny.destroy: Create destroyable modules in 'Shiny'}
\description{
A way to fully remove dynamic 'Shiny' modules from a session.
}
\author{
\strong{Maintainer}: Ashley Baldry \email{arbaldry91@gmail.com}

}
