% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.destroy.R
\docType{package}
\name{shiny.destroy-package}
\alias{shiny.destroy}
\alias{shiny.destroy-package}
\title{shiny.destroy: Use dynamic, destroyable modules in 'Shiny'}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
